requirements.txt
a2wsgi==1.10.8
aiohappyeyeballs==2.6.1
aiohttp==3.10.11
aiologic==0.14.0
aiosignal==1.3.2
aiosqlite==0.21.0
alembic==1.16.1
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
apache-airflow==2.8.0
apache-airflow-core==3.0.1
apache-airflow-providers-common-compat==1.7.0
apache-airflow-providers-common-io==1.6.0
apache-airflow-providers-common-sql==1.20.0
apache-airflow-providers-ftp==3.11.1
apache-airflow-providers-http==4.13.3
apache-airflow-providers-imap==3.7.0
apache-airflow-providers-smtp==2.1.0
apache-airflow-providers-sqlite==3.9.1
apache-airflow-providers-standard==1.2.0
apache-airflow-task-sdk==1.0.1
apispec==6.8.2
argcomplete==3.6.2
asgiref==3.8.1
async-timeout==5.0.1
attrs==25.3.0
babel==2.17.0
backports.strenum==1.3.1
beautifulsoup4==4.13.4
blinker==1.9.0
blis==1.3.0
branca==0.8.1
bs4==0.0.2
cachelib==0.13.0
cachetools==5.5.2
cadwyn==5.3.3
catalogue==2.0.10
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
clickclick==20.10.2
cloudpathlib==0.21.1
colorama==0.4.6
colorlog==4.8.0
confection==0.1.5
ConfigUpdater==3.2
confluent-kafka==2.10.0
connexion==2.14.2
cron-descriptor==1.4.5
croniter==6.0.0
cryptography==45.0.3
cymem==2.0.11
de_core_news_sm @ https://github.com/explosion/spacy-models/releases/download/de_core_news_sm-3.8.0/de_core_news_sm-3.8.0-py3-none-any.whl
delta-spark==3.3.2
Deprecated==1.2.18
dill==0.4.0
dnspython==2.7.0
email-validator==1.3.1
exceptiongroup==1.3.0
fastapi==0.115.12
fastapi-cli==0.0.7
feedparser==6.0.11
filelock==3.18.0
Flask==2.2.5
Flask-AppBuilder==4.3.10
Flask-Babel==2.0.0
Flask-Caching==2.3.1
Flask-JWT-Extended==4.7.1
Flask-Limiter==3.12
Flask-Login==0.6.3
Flask-Session==0.8.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.2.2
folium==0.19.6
frozenlist==1.6.0
fsspec==2025.5.1
geographiclib==2.0
geopy==2.4.1
gitdb==4.0.12
GitPython==3.1.44
google-re2==1.1.20240702
googleapis-common-protos==1.70.0
graphviz==0.20.3
greenlet==3.2.2
grpcio==1.72.1
gunicorn==23.0.0
h11==0.16.0
hf-xet==1.1.2
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.32.2
idna==3.10
importlib_metadata==8.6.1
inflection==0.5.1
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
kafka-python==2.2.10
langcodes==3.5.0
language_data==1.3.0
lazy-object-proxy==1.11.0
libcst==1.8.0
limits==5.2.0
linkify-it-py==2.0.3
lockfile==0.12.2
lxml==5.4.0
Mako==1.3.10
marisa-trie==1.2.1
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
marshmallow-oneofschema==3.2.0
marshmallow-sqlalchemy==0.26.1
mdit-py-plugins==0.4.2
mdurl==0.1.2
methodtools==0.4.7
more-itertools==10.7.0
mpmath==1.3.0
msgspec==0.19.0
multidict==6.4.4
murmurhash==1.0.13
narwhals==1.41.0
networkx==3.4.2
numpy==2.2.6
opentelemetry-api==1.33.1
opentelemetry-exporter-otlp==1.33.1
opentelemetry-exporter-otlp-proto-common==1.33.1
opentelemetry-exporter-otlp-proto-grpc==1.33.1
opentelemetry-exporter-otlp-proto-http==1.33.1
opentelemetry-proto==1.33.1
opentelemetry-sdk==1.33.1
opentelemetry-semantic-conventions==0.54b1
ordered-set==4.1.0
packaging==24.2
pandas==2.2.3
pathspec==0.12.1
pendulum==2.1.2
pillow==11.2.1
pluggy==1.6.0
praw==7.8.1
prawcore==2.4.0
preshed==3.0.10
prison==0.2.1
propcache==0.3.1
protobuf==5.29.5
psutil==7.0.0
py4j==0.10.9.7
pyarrow==20.0.0
pycparser==2.22
pydantic==2.11.5
pydantic_core==2.33.2
pydeck==0.9.1
Pygments==2.19.1
PyJWT==2.10.1
pymongo==4.13.0
pyspark==3.5.6
python-daemon==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
python-nvd3==0.16.0
python-slugify==8.0.4
pytz==2025.2
pytzdata==2020.1
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
retryhttp==1.3.3
rfc3339-validator==0.1.4
rich==13.9.4
rich-argparse==1.7.1
rich-toolkit==0.14.7
rpds-py==0.25.1
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.3
sentence-transformers==4.1.0
setproctitle==1.3.6
sgmllib3k==1.0.0
shellingham==1.5.4
six==1.17.0
smart-open==7.1.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.7
spacy==3.8.7
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==1.4.54
SQLAlchemy-JSONField==1.0.2
SQLAlchemy-Utils==0.41.2
sqlparse==0.5.3
srsly==2.5.1
starlette==0.46.2
streamlit==1.45.1
streamlit_folium==0.25.0
structlog==25.4.0
svcs==25.1.0
sympy==1.14.0
tabulate==0.9.0
tenacity==9.1.2
termcolor==3.1.0
text-unidecode==1.3
thinc==8.3.6
threadpoolctl==3.6.0
tokenizers==0.21.1
toml==0.10.2
tomli==2.2.1
torch==2.7.0
tornado==6.5.1
tqdm==4.67.1
transformers==4.52.3
typer==0.16.0
types-requests==2.32.0.20250602
typing-inspection==0.4.1
typing_extensions==4.13.2
tzdata==2025.2
uc-micro-py==1.0.3
unicodecsv==0.14.1
universal_pathlib==0.2.6
update-checker==0.18.0
urllib3==2.4.0
uuid6==2024.7.10
uvicorn==0.34.3
uvloop==0.21.0
wasabi==1.1.3
watchdog==6.0.0
watchfiles==1.0.5
weasel==0.4.1
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==2.2.3
wirerope==1.0.0
wrapt==1.17.2
WTForms==3.2.1
xyzservices==2025.4.0
yarl==1.20.0
zipp==3.22.0
